requirements.txt
torch         == 1.13.1+cu117
torchvision   == 0.14.1+cu117
numpy         == 1.24.3
scipy         == 1.9.1
pandas        == 2.0.1
matplotlib    == 3.7.1
Pillow        == 9.5.0
scikit-image  == 0.20.0
opencv-python == 4.7.0.72
ipython       == 8.13.0
plotly        == 5.14.1

